+++ Heartland-iOS-SDK/Classes/Terminals/PAX/Models/HpsPaxCreditResponse.m
@@ -15,7 +15,7 @@
 - (HpsBinaryDataScanner*) parseResponse{
     HpsBinaryDataScanner *reader = [super parseResponse];
     //must parse responses for duplicate error codes (100011 - Local Duplicate, 103000 Host Duplicate)
-    if ([self.deviceResponseCode isEqualToString:@"000000"] || [self.deviceResponseCode isEqualToString:@"100011"]  || [self.deviceResponseCode isEqualToString:@"103000"]) {
+    if ([self.deviceResponseCode isEqualToString:@"000000"] || [self.deviceResponseCode isEqualToString:@"100011"]  || [self.deviceResponseCode isEqualToString:@"103000"] || [self.deviceResponseCode isEqualToString:@"000002"]) {
         
         self.hostResponse = [[HpsPaxHostResponse alloc] initWithBinaryReader:reader];
         self.transactionType = [reader readStringUntilDelimiter:HpsControlCodes_FS];