+++ Heartland-iOS-SDK/Classes/Terminals/PAX/Models/HpsPaxDeviceResponse.m
@@ -74,6 +74,7 @@
 
 
 			self.cardBin = [self.extDataResponse.collection objectForKey:PAX_EXT_DATA_CARD_BIN];
+            self.programType = [self.extDataResponse.collection objectForKey:PAX_EXT_DATA_PROGRAM_TYPE];
 			self.signatureStatus = [self.extDataResponse.collection objectForKey:PAX_EXT_DATA_SIGNATURE_STATUS];
             self.pinEntryStatus = [self.extDataResponse.collection objectForKey:PAX_EXT_DATA_PIN_ENTRY_STATUS];
             self.printLine1 = [self.extDataResponse.collection objectForKey:PAX_EXT_DATA_PRINT_LINE_1];