+++ Heartland-iOS-SDK/Classes/Terminals/UPA/Models/HpsUpaResponse.h
@@ -46,6 +46,7 @@
 @property (nonatomic,strong) NSString *response;
 @property (nonatomic,strong) NSString *multipleMessage;
 @property (nonatomic,strong) NSString *resultText;
+@property (nonatomic,strong) NSString *responseText;
 @property (nonatomic,strong) NSString *responseCode;
 @property (nonatomic,strong) NSString *cardAcquisition;
 @property (nonatomic,strong) NSString *signatureLine;