+++ Heartland-iOS-SDK/Classes/Terminals/Utility/GlobalMobileSDK/GMSResponseHelper.swift
@@ -62,6 +62,7 @@ class GMSResponseHelper {
         data.terminalVerficationResult = resp?.tvr
         data.cardType = HpsC2xEnums.cardTypeToString(resp?.cardType)
         data.authCode = resp?.authCode
+        data.issuerRspMsg = resp?.emvIssuerResponse
         // data.authCodeData = resp?.authCodeData
 
         if let uintValue = resp?.tip {
@@ -164,6 +165,7 @@ class GMSResponseHelper {
         data.entryMode = Int32(HpsC2xEnums.cardDataSourceTypeToEntryMode(resp?.cardDataSourceType).rawValue)
         data.approvalCode = resp?.authCode
         data.maskedCardNumber = resp?.maskedPan
+        data.responseCode = resp?.gatewayResponseCode
         data.responseText = resp?.gatewayResponseText
         data.deviceResponseCode = deviceResponseCode
         data.terminalRefNumber = resp?.posReferenceNumber
@@ -175,7 +177,7 @@ class GMSResponseHelper {
         data.terminalVerficationResult = resp?.tvr
         data.cardType = HpsC2xEnums.cardTypeToString(resp?.cardType)
         data.issuerRspCode = resp?.emvIssuerRspCode
-        data.issuerRspMsg = resp?.emvIssuerResponse
+        data.issuerRspMsg = resp?.hostProcessingResult?.emvIssuerAuthenticationData
         data.authCode = resp?.authCode
         data.authCodeData = resp?.authCodeData
         if let uintValue = resp?.tip {