+++ Heartland-iOS-SDK/Libraries/GlobalPaymentsApi.xcframework/ios-arm64/GlobalPaymentsApi.framework/Headers/GPTransactionReference.h
@@ -1,0 +1,17 @@
+#ifndef GPTransactionReference_h
+#define GPTransactionReference_h
+
+#import <Foundation/Foundation.h>
+#import <GlobalPaymentsApi/GPPaymentMethod.h>
+#import <GlobalPaymentsApi/GPEnums.h>
+
+@interface GPTransactionReference : GPPaymentMethod
+
+@property (nonatomic, strong) NSString* authCode;
+@property (nonatomic, strong) NSString* batchNumber;
+@property (nonatomic, strong) NSString* transactionId;
+@property (nonatomic, strong) NSString* clientTransactionId;
+
+@end
+
+#endif /* GPTransactionReference_h */